# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DentalEase API",
        "description": "API for DentalEase application - managing dental services and user authentication",
        "contact": {
            "name": "DentalEase Team",
            "email": "support@dentalease.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Development server"
        }
    ],
    "paths": {
        "/appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get all appointments",
                "description": "Retrieve a list of all appointments",
                "operationId": "e975113000704ff889fd39adcceb36bf",
                "responses": {
                    "200": {
                        "description": "List of appointments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Appointment"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Create a new appointment",
                "description": "Create a new appointment",
                "operationId": "f2581d2b5d5ab8807ceb65d47549cf1a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_name",
                                    "patient_email",
                                    "patient_contact",
                                    "purpose",
                                    "schedule_time",
                                    "schedule_date",
                                    "status"
                                ],
                                "properties": {
                                    "patient_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "patient_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "patient_contact": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "purpose": {
                                        "type": "string",
                                        "example": "Regular dental checkup"
                                    },
                                    "remarks": {
                                        "type": "string",
                                        "example": "Patient has sensitive teeth"
                                    },
                                    "schedule_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "14:30"
                                    },
                                    "schedule_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-15"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "active",
                                            "past_due",
                                            "cancelled"
                                        ],
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Appointment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/appointments/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get a specific appointment",
                "description": "Retrieve a specific appointment by ID",
                "operationId": "62e6b32be0a1e6a7fab0708aef3e02fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Appointment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Appointment not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Update an appointment",
                "description": "Update an existing appointment",
                "operationId": "df9efe502863aad3ea47cc918833aaa5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "patient_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "patient_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "patient_contact": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "purpose": {
                                        "type": "string",
                                        "example": "Regular dental checkup"
                                    },
                                    "remarks": {
                                        "type": "string",
                                        "example": "Patient has sensitive teeth"
                                    },
                                    "schedule_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "14:30"
                                    },
                                    "schedule_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-15"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "active",
                                            "past_due",
                                            "cancelled"
                                        ],
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Appointment updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Appointment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Appointment not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Delete an appointment",
                "description": "Delete an existing appointment",
                "operationId": "6b486c1ca5a3fb49659004865f595ca0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Appointment deleted successfully"
                    },
                    "404": {
                        "description": "Appointment not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Appointment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "patient_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "patient_email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "patient_contact": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "purpose": {
                        "type": "string",
                        "example": "Regular dental checkup"
                    },
                    "remarks": {
                        "type": "string",
                        "example": "Patient has sensitive teeth"
                    },
                    "schedule_time": {
                        "type": "string",
                        "format": "time",
                        "example": "14:30"
                    },
                    "schedule_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-08-15"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "active",
                            "past_due",
                            "cancelled"
                        ],
                        "example": "pending"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-31T09:10:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-31T09:10:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}